backend/requirements.txt
africastalking==2.0.2
altair==6.0.0
amqp==5.3.1
APScheduler==3.11.2
asgiref==3.11.1
attrs==26.1.0
autobahn==25.12.2
Automat==25.4.16
bcrypt==5.0.0
billiard==4.2.4
blinker==1.9.0
boto3==1.43.0
botocore==1.43.0
cachetools==7.0.5
captcha==0.7.1
cbor2==6.0.1
celery==5.6.3
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.7
click==8.3.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.3
cryptography==46.0.7
cycler==0.12.1
daphne==4.2.1
defusedxml==0.7.1
dj-database-url==3.1.2
dj-rest-auth==7.2.0
Django==6.0.4
django-allauth==65.16.1
django-allauth-2fa==0.11.1
django-channels==0.7.0
django-cors-headers==4.9.0
django-filter==25.2
django-otp==1.7.0
django-redis==6.0.0
django-rest-knox==5.0.4
django-storages==1.14.6
django-tenants==3.10.1
djangochannelsrestframework==1.3.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
et_xmlfile==2.0.0
extra-streamlit-components==0.1.81
Flask==3.1.3
fonttools==4.62.1
fpdf2==2.8.7
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.4.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
inflection==0.5.1
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
kombu==5.6.2
MarkupSafe==3.0.3
matplotlib==3.10.8
msgpack==1.1.2
narwhals==2.19.0
numpy==2.4.4
oauthlib==3.3.1
openpyxl==3.1.5
packaging==26.2
pandas==3.0.2
pillow==12.2.0
plotly==6.7.0
pluggy==1.6.0
prompt_toolkit==3.0.52
protobuf==7.34.1
psycopg2-binary==2.9.12
py-ubjson==0.16.1
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydeck==0.9.1
Pygments==2.20.0
PyJWT==2.12.1
pyOpenSSL==26.1.0
pyparsing==3.3.2
pytest==9.0.3
pytest-django==4.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
qrcode==8.2
redis==7.4.0
referencing==0.37.0
reportlab==4.5.0
requests==2.33.1
requests-oauthlib==2.0.0
responses==0.26.0
rpds-py==0.30.0
s3transfer==0.17.0
schema==0.7.8
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
service-identity==24.2.0
six==1.17.0
smmap==5.0.3
SQLAlchemy==2.0.49
sqlparse==0.5.5
streamlit==1.56.0
streamlit-authenticator==0.4.2
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.5
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
ujson==5.12.0
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
watchdog==6.0.0
wcwidth==0.6.0
Werkzeug==3.1.8
whitenoise==6.12.0
zope.interface==8.4
